Fix prover lookup
